heist is working

+++ Models/Muscle.cs
@@ -6,7 +6,7 @@ namespace heist_two
   {
     public string Name { get; set; }
     public int SkillLevel { get; set; }
-    public int PercentageCut { get; set; }
+    public double PercentageCut { get; set; }
     public void PerformSkill(Bank Bank)
     {
       Bank.SecurityGuardScore -= SkillLevel;
